+++ src/garminhd/GarminHDReceive.cpp
@@ -490,7 +490,7 @@ bool GarminHDReceive::UpdateScannerStatus(int status) {
 }
 
 bool GarminHDReceive::ProcessReport(const uint8_t *report, size_t len) {
-  LOG_BINARY_RECEIVE(wxT("ProcessReport"), report, len);
+  LOG_BINARY_REPORTS(wxString::Format(wxT("%s report"), m_ri->m_name.c_str()), report, len);
 
   time_t now = time(0);